BAH-4911 | Refactor. Restruture Patient Search Files
