+++ packages/bahmni-widgets/src/search/searchPatient/strategies/NameOrIdSearchStrategy.ts
@@ -8,7 +8,7 @@ import {
   SearchStrategy,
   SearchContext,
   ValidationResult,
-} from '../SearchStrategy.interface';
+} from './SearchStrategy.interface';
 
 /**
  * Strategy for searching patients by name or patient ID
