+++ packages/bahmni-widgets/src/search/searchPatient/strategies/SearchStrategy.interface.ts
